Rename the dbt_util to dbtutil
